requirements.txt
aiogram==2.24
matplotlib==3.3.4
nest_asyncio==1.5.1
Pillow==9.4.0
torch==1.11.0
torchvision==0.12.0
